feat: 유저 자산 집계 서비스에 가상자산 가격 변환 집계 적용

+++ applications/api-server/src/main/java/com/gamzabit/api/user/service/assets/vo/AggregatedUserAsset.java
@@ -1,10 +1,10 @@
 package com.gamzabit.api.user.service.assets.vo;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.List;
 
 public record AggregatedUserAsset(
-    BigInteger totalUserAssetKrw,
+    BigDecimal totalUserAssetKrw,
     List<UserAsset> assets
 ) {
 }
