+++ applications/api-server/src/main/java/com/gamzabit/api/user/service/assets/vo/UserAssetWithKrw.java
@@ -1,9 +1,9 @@
 package com.gamzabit.api.user.service.assets.vo;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public record UserAssetWithKrw(
     UserAsset userAsset,
-    BigInteger krw
+    BigDecimal krw
 ) {
 }
